Refactor OptimizedGenericIterativeAlgorithm at least a bit T_T
